fix(query): only rename concepts
