+++ rust/dialog-macros/src/lib.rs
@@ -362,7 +362,7 @@ pub fn derive_formula(input: TokenStream) -> TokenStream {
 /// // Create attribute values
 /// let name = employee::Name("Alice".to_string());
 /// ```
-#[proc_macro_derive(Attribute, attributes(cardinality, domain, namespace, dialog))]
+#[proc_macro_derive(Attribute, attributes(cardinality, domain, namespace))]
 pub fn derive_attribute(input: TokenStream) -> TokenStream {
     query::attribute::derive(input)
 }
